### 2025-04-11 418e203
ResNet parametrized and it seems all bugs resolved
