### 2025-04-11 6c6f524
vggnetparametric test file created
