### 2025-04-11 348cda5
vggnetparametric test file created
