+++ temp/CMakeLists.txt
@@ -1332,3 +1332,43 @@ target_link_libraries(${PROJECT_NAME}
 
 
 
+####################################### VGGNETPARAMETRIC #######################################
+
+project(vggnetparametric)
+list(APPEND CMAKE_PREFIX_PATH /home/kami/libs/c++/libtorch/)
+find_package(Torch REQUIRED)
+#find_package(xTorch REQUIRED)
+find_package(OpenCV REQUIRED)
+find_package(LibLZMA REQUIRED)
+find_package(LibArchive REQUIRED)
+
+set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD_REQUIRED True)
+
+set(TEST_SOURCE_FILES ${UTILITY_FILES} ${MEDIA_FILES} ${DEFINITION_FILES} ${DATASET_FILES} ${MODEL_FILES}  ${BASE_SOURCE_FILES}   vggnetparametric.cpp)
+
+add_executable(${PROJECT_NAME} ${TEST_SOURCE_FILES})
+target_link_libraries(${PROJECT_NAME} "${TORCH_LIBRARIES}")
+#
+set_property(TARGET ${PROJECT_NAME} PROPERTY CXX_STANDARD 17)
+
+include_directories(/usr/include)
+
+target_include_directories(${PROJECT_NAME} PRIVATE ${LIBTAR_INCLUDE_DIR} ${TORCH_INCLUDE_DIRS} ${LibArchive_INCLUDE_DIRS})
+target_link_libraries(${PROJECT_NAME}
+        ${CURL_LIBRARIES}
+        ${TORCH_LIBRARIES}
+        ${OpenCV_LIBS}
+        ${ZLIB_LIBS}
+        ${ZIP_LIBRARIES}
+        zip
+        ${ZLIB_LIBRARIES}
+        ${LIBTAR_LIBRARY}
+        OpenSSL::SSL
+        OpenSSL::Crypto
+        LibLZMA::LibLZMA
+        ${LibArchive_LIBRARIES}
+)
+
+
+
