### 2025-04-11 78e0660
alexnetparametric test file created
