### 2025-04-11 6e437e7
alexnetparametric test file created
