### 2025-04-11 2a5823a
resnetparametric test file created
