### 2025-04-11 ced91c8
resnetparametric test file created and worked properly
