### 2025-04-11 c1be099
HMDB51 and ImageNet first constructor marked as explicit
