+++ include/datasets/specific/imagenet.h
@@ -6,7 +6,7 @@
 namespace xt::data::datasets {
     class ImageNet : BaseDataset {
     public :
-        ImageNet(const std::string &root);
+        explicit  ImageNet(const std::string &root);
         ImageNet(const std::string &root, DataMode mode);
         ImageNet(const std::string &root, DataMode mode , bool download);
         ImageNet(const std::string &root, DataMode mode , bool download, TransformType transforms);
