### 2025-04-11 1deb16d
HMDB51 and ImageNet first constructor marked as explicit
