### 2025-04-11 b66bed9
ImageNette and InStereo2k first constructor marked as explicit
