### 2025-04-11 79688fe
ImageNette and InStereo2k first constructor marked as explicit
