### 2025-04-11 edb580b
INaturalist and Kinetics first constructor marked as explicit
