### 2025-04-11 d6ebacf
lfw datasets first constructor marked as explicit
