+++ include/datasets/specific/lfw.h
@@ -9,7 +9,7 @@ namespace fs = std::filesystem;
 namespace xt::data::datasets {
     class LFW : public BaseDataset {
     public :
-        LFW(const std::string &root);
+        explicit LFW(const std::string &root);
         LFW(const std::string &root, DataMode mode);
         LFW(const std::string &root, DataMode mode , bool download);
         LFW(const std::string &root, DataMode mode , bool download, TransformType transforms);
@@ -72,7 +72,7 @@ namespace xt::data::datasets {
 
          */
     public :
-        LFWPeople(const std::string &root);
+        explicit LFWPeople(const std::string &root);
         LFWPeople(const std::string &root, DataMode mode);
         LFWPeople(const std::string &root, DataMode mode , bool download);
         LFWPeople(const std::string &root, DataMode mode , bool download, TransformType transforms);
@@ -108,7 +108,7 @@ namespace xt::data::datasets {
 
          */
     public :
-        LFWPairs(const std::string &root);
+        explicit  LFWPairs(const std::string &root);
         LFWPairs(const std::string &root, DataMode mode);
         LFWPairs(const std::string &root, DataMode mode , bool download);
         LFWPairs(const std::string &root, DataMode mode , bool download, TransformType transforms);
