### 2025-04-11 3deed4f
mnist datasets first constructor marked as explicit
