### 2025-04-11 2cc3085
moving-mnist and omniglot datasets first constructor marked as explicit
