### 2025-04-11 3a8af20
oxford,pcam,photo-tours and places datasets first constructor marked as explicit
