### 2025-04-11 51ed283
oxford,pcam,photo-tours and places datasets first constructor marked as explicit
