### 2025-04-11 ec3e6c8
stanford-cars,stl,sun datasets first constructor marked as explicit
