### 2025-04-11 3d14fc0
svhn,ucf,usps datasets first constructor marked as explicit
