### 2025-04-11 c7ba5a7
svhn,ucf,usps datasets first constructor marked as explicit
