### 2025-04-11 c3a71c6
voc and widerface datasets first constructor marked as explicit
