+++ include/models/cnn/unet.h
@@ -9,61 +9,60 @@
 #include "../../exceptions/implementation.h"
 
 namespace xt::models {
-
-
-//class DoubleConv(nn.Module):
-//    def __init__(self, in_channels, out_channels):
-//        super().__init__()
-//        self.conv_op = nn.Sequential(
-//                nn.Conv2d(in_channels, out_channels, kernel_size=3, padding=1),
-//                nn.ReLU(inplace=True),
-//                nn.Conv2d(out_channels, out_channels, kernel_size=3, padding=1),
-//                nn.ReLU(inplace=True)
-//        )
-//
-//    def forward(self, x):
-//        return self.conv_op(x)
+    //class DoubleConv(nn.Module):
+    //    def __init__(self, in_channels, out_channels):
+    //        super().__init__()
+    //        self.conv_op = nn.Sequential(
+    //                nn.Conv2d(in_channels, out_channels, kernel_size=3, padding=1),
+    //                nn.ReLU(inplace=True),
+    //                nn.Conv2d(out_channels, out_channels, kernel_size=3, padding=1),
+    //                nn.ReLU(inplace=True)
+    //        )
+    //
+    //    def forward(self, x):
+    //        return self.conv_op(x)
 
 
     struct DoubleConv : torch::nn::Module {
     public:
         DoubleConv(int in_channels, int out_channels);
+
         torch::Tensor forward(torch::Tensor input);
+
     private:
         torch::nn::Sequential conv_op = nullptr;
-
-
     };
 
 
-//class DownSample(nn.Module):
-//    def __init__(self, in_channels, out_channels):
-//        super().__init__()
-//        self.conv = DoubleConv(in_channels, out_channels)
-//        self.pool = nn.MaxPool2d(kernel_size=2, stride=2)
-//
-//    def forward(self, x):
-//        down = self.conv(x)
-//        p = self.pool(down)
-//        return down, p
+    //class DownSample(nn.Module):
+    //    def __init__(self, in_channels, out_channels):
+    //        super().__init__()
+    //        self.conv = DoubleConv(in_channels, out_channels)
+    //        self.pool = nn.MaxPool2d(kernel_size=2, stride=2)
+    //
+    //    def forward(self, x):
+    //        down = self.conv(x)
+    //        p = self.pool(down)
+    //        return down, p
 
     struct DownSample : torch::nn::Module {
     public:
         DownSample(int in_channels, int out_channels);
+
         torch::Tensor forward(torch::Tensor input);
+
     private:
-        DoubleConv conv{0,0} ;
+        DoubleConv conv{0, 0};
         torch::nn::MaxPool2d pool = nullptr;
     };
 
 
-
-//class UpSample(nn.Module):
-//    def __init__(self, in_channels, out_channels):
-//        super().__init__()
-//        self.up = nn.ConvTranspose2d(in_channels, in_channels//2, kernel_size=2, stride=2)
-//        self.conv = DoubleConv(in_channels, out_channels)
-//
+    //class UpSample(nn.Module):
+    //    def __init__(self, in_channels, out_channels):
+    //        super().__init__()
+    //        self.up = nn.ConvTranspose2d(in_channels, in_channels//2, kernel_size=2, stride=2)
+    //        self.conv = DoubleConv(in_channels, out_channels)
+    //
     // def forward(self, x1, x2):
     //     x1 = self.up(x1)
     //     x = torch.cat([x1, x2], 1)
@@ -72,52 +71,70 @@ namespace xt::models {
     struct UpSample : torch::nn::Module {
     public:
         UpSample(int in_channels, int out_channels);
[…]
