### 2025-04-12 ffd67a9
MobileNetV3 interface created
