+++ include/models/cnn/mobilenet.h
@@ -8,176 +8,208 @@
 
 
 namespace xt::models {
+    //class HSigmoid(nn.Module):
+    //    def __init__(self):
+    //        """Hard Sigmoid activation function."""
+    //        super().__init__()
+    //        self.relu6 = nn.ReLU6(inplace=True)
+    //
+    //    def forward(self, x):
+    //        x = self.relu6(x + 3) / 6
+    //        return x
+    struct HSigmoid : torch::nn::Module {
+    public:
+        HSigmoid();
+
+        torch::Tensor forward(torch::Tensor x);
+
+    private:
+        torch::nn::ReLU6 relu6;
+    };
+
+    //---------------------------------------------------------------------------
+
+    //class HSwish(nn.Module):
+    //    def __init__(self):
+    //        """Hard Swish activation function."""
+    //        super().__init__()
+    //        self.relu6 = nn.ReLU6(inplace=True)
+    //
+    //    def forward(self, x):
+    //        x = x * self.relu6(x + 3) / 6
+    //        return x
+    //
+    //
+    struct HSwish : torch::nn::Module {
+    public:
+        HSwish();
+
+        torch::Tensor forward(torch::Tensor x);
+
+    private:
+        torch::nn::ReLU6 relu6;
+    };
+
+    //---------------------------------------------------------------------------
+
+    //class SqueezeExcite(nn.Module):
+    //    def __init__(self,input_channels: int,squeeze: int = 4,) -> None:
+    //        """
+    //        Squeeze-and-Excitation block.
+    //
+    //        Args:
+    //        input_channels (`int`): Number of input channels.
+    //        squeeze (`int`, optional): Squeeze ratio. Defaults to 4.
+    //        """
+    //        super().__init__()
+    //
+    //        self.SE = nn.Sequential(
+    //                nn.AdaptiveAvgPool2d(output_size=1),
+    //                nn.Conv2d(input_channels, out_channels=input_channels//squeeze, kernel_size=1, stride=1, bias=False),
+    //        nn.BatchNorm2d(input_channels//squeeze),
+    //        nn.ReLU(inplace=True),
+    //        nn.Conv2d(input_channels//squeeze, input_channels, kernel_size=1, stride=1, bias=False),
+    //        nn.BatchNorm2d(input_channels),
+    //                HSigmoid(),
+    //        )
+    //
+    //    def forward(self, x):
+    //        x = x * self.SE(x)
+    //        return x
+
+    struct SqueezeExcite : torch::nn::Module {
+    public:
+        SqueezeExcite(int input_channels, int squeeze = 4);
+
+        torch::Tensor forward(torch::Tensor x);
+
+    private:
+        torch::nn::Sequential SE;
+    };
 
 
-//class SqueezeExcite(nn.Module):
-//    def __init__(self,input_channels: int,squeeze: int = 4,) -> None:
-//        """
-//        Squeeze-and-Excitation block.
-//
-//        Args:
-//        input_channels (`int`): Number of input channels.
-//        squeeze (`int`, optional): Squeeze ratio. Defaults to 4.
-//        """
-//        super().__init__()
-//
-//        self.SE = nn.Sequential(
-//                nn.AdaptiveAvgPool2d(output_size=1),
-//                nn.Conv2d(input_channels, out_channels=input_channels//squeeze, kernel_size=1, stride=1, bias=False),
-//        nn.BatchNorm2d(input_channels//squeeze),
-//        nn.ReLU(inplace=True),
-//        nn.Conv2d(input_channels//squeeze, input_channels, kernel_size=1, stride=1, bias=False),
-//        nn.BatchNorm2d(input_channels),
-//                HSigmoid(),
-//        )
-//
-//    def forward(self, x):
[…]
