### 2025-04-12 a72ab9f
MobileNetV3 interface created
