+++ include/models/cnn/mobilenet.h
@@ -146,11 +146,14 @@ namespace xt::models {
 
     struct Bottleneck : torch::nn::Module {
     public:
-        Bottleneck(int input_channels,int kernel,int stride,int expansion,int output_channels,torch::nn::Module activation ,bool se = false);
+        Bottleneck(int input_channels, int kernel, int stride, int expansion, int output_channels,
+                   torch::nn::Module activation, bool se = false);
+
         torch::Tensor forward(torch::Tensor x);
 
     private:
-
+        torch::nn::Sequential bottleneck;
+        torch::nn::Sequential downsample;
     };
 
 
