+++ include/models/cnn/mobilenet.h
@@ -223,5 +223,15 @@ namespace xt::models {
     struct MobileNetV3 : BaseModel {
     public:
         MobileNetV3();
+
+        MobileNetV3(int input_channels, int num_classes, float dropout_prob = 0.5);
+        torch::Tensor forward(torch::Tensor x);
+
+    private:
+        torch::nn::Sequential initial_conv;
+        torch::nn::Sequential bottlenecks;
+        torch::nn::Sequential final_conv;
+        torch::nn::AdaptiveAvgPool2d pool;
+        torch::nn::Sequential classifier;
     };
 }
