+++ src/models/cnn/mobilenet.cpp
@@ -9,7 +9,24 @@ namespace xt::models {
         x = this->relu6(x + 3) / 6;
         return x;
     }
+
     // --------------------------------------------------------------------
+
+    HSwish::HSwish() {
+        this->relu6 = torch::nn::ReLU6();
+    }
+
+    torch::Tensor HSwish::forward(torch::Tensor x) {
+        x = this->relu6(x + 3) / 6;
+        return x;
+    }
+
+    // --------------------------------------------------------------------
+
+    SqueezeExcite::SqueezeExcite(int input_channels, int squeeze) {
+
+
+    }
 
 
     MobileNetV3::MobileNetV3() {
