+++ include/models/cnn/mobilenet.h
@@ -222,8 +222,6 @@ namespace xt::models {
 
     struct MobileNetV3 : BaseModel {
     public:
-        MobileNetV3();
-
         MobileNetV3(int input_channels, int num_classes, float dropout_prob = 0.5);
         torch::Tensor forward(torch::Tensor x);
 
