+++ src/models/cnn/mobilenet.cpp
@@ -65,11 +65,11 @@ namespace xt::models {
 
         this->bottlenecks = torch::nn::Sequential(
             Bottleneck(16, 3, 1, 16, 16, static_cast<Module>(*torch::nn::ReLU(torch::nn::ReLUOptions(true)))),
-            Bottleneck(16, 3, 2, 64, 24, *torch::nn::ReLU(torch::nn::ReLUOptions(true))),
-            Bottleneck(24, 3, 1, 72, 24, *torch::nn::ReLU(torch::nn::ReLUOptions(true))),
-            Bottleneck(24, 5, 2, 72, 40, *torch::nn::ReLU(torch::nn::ReLUOptions(true))),
-            Bottleneck(40, 5, 1, 120, 40, *torch::nn::ReLU(torch::nn::ReLUOptions(true))),
-            Bottleneck(40, 5, 1, 120, 40, *torch::nn::ReLU(torch::nn::ReLUOptions(true))),
+            Bottleneck(16, 3, 2, 64, 24, static_cast<Module>(*torch::nn::ReLU(torch::nn::ReLUOptions(true)))),
+            Bottleneck(24, 3, 1, 72, 24, static_cast<Module>(*torch::nn::ReLU(torch::nn::ReLUOptions(true)))),
+            Bottleneck(24, 5, 2, 72, 40, static_cast<Module>(*torch::nn::ReLU(torch::nn::ReLUOptions(true)))),
+            Bottleneck(40, 5, 1, 120, 40, static_cast<Module>(*torch::nn::ReLU(torch::nn::ReLUOptions(true)))),
+            Bottleneck(40, 5, 1, 120, 40, static_cast<Module>(*torch::nn::ReLU(torch::nn::ReLUOptions(true)))),
             Bottleneck(40, 3, 2, 240, 80, static_cast<Module>(HSwish())),
             Bottleneck(80, 3, 1, 200, 80, static_cast<Module>(HSwish())),
             Bottleneck(80, 3, 1, 184, 80, static_cast<Module>(HSwish())),
