### 2025-04-14 da161e8
standalone examples moved to another repository
