### 2025-04-14 b7b0b9b
standalone examples moved to another repository
