+++ examples-wil-be-moved-to-another-repo/models/resnet.cpp
@@ -1,366 +1,0 @@
-#include <torch/torch.h>
-#include <torch/data/datasets/mnist.h>
-#include <iostream>
-#include <vector>
-#include <fstream>
-#include <string>
-
-#define DEBUG_MODE true
-
-using namespace std;
-
-class CIFAR10 : public torch::data::Dataset<CIFAR10> {
-public:
-    CIFAR10(const std::string &root) {
-        // Load data from the specified root directory
-        load_data(root);
-    }
-
-    // Override the get method to return a sample
-    torch::data::Example<> get(size_t index) override {
-        // Return the tensor image and its corresponding label
-        return {data[index].clone(), torch::tensor(labels[index])}; // Clone to ensure tensor validity
-    }
-
-    // Override the size method to return the number of samples
-    torch::optional<size_t> size() const override {
-        return data.size();
-    }
-
-private:
-    std::vector<torch::Tensor> data; // Store image data as tensors
-    std::vector<int64_t> labels;      // Store labels
-
-    void load_data(const std::string &root) {
-        const int num_files = 5;
-        for (int i = 1; i <= num_files; ++i) {
-            std::string file_path = root + "/data_batch_" + std::to_string(i) + ".bin";
-            std::ifstream file(file_path, std::ios::binary);
-            if (!file.is_open()) {
-                std::cerr << "Failed to open file: " << file_path << std::endl;
-                continue;
-            }
-
-            for (int j = 0; j < 10000; ++j) {
-                uint8_t label;
-                file.read(reinterpret_cast<char *>(&label), sizeof(label));
-                labels.push_back(static_cast<int64_t>(label));
-
-                std::vector<uint8_t> image(3072); // 32x32x3 = 3072
-                file.read(reinterpret_cast<char *>(image.data()), image.size());
-
-                // Reshape the image to 3x32x32 and convert to a Torch tensor
-                auto tensor_image = torch::from_blob(image.data(), {3, 32, 32},
-                                                     torch::kByte).clone(); // Clone to ensure memory management
-                tensor_image = tensor_image.permute({0, 2, 1}); // Permute to get the correct order (C, H, W)
-
-                data.push_back(tensor_image); // Store the tensor in the data vector
-            }
-
-            file.close();
-        }
-    }
-};
-
-//class ResidualBlock(nn.Module):
-//    def __init__(self, in_channels, out_channels, stride = 1, downsample = None):
-//        super(ResidualBlock, self).__init__()
-//        self.conv1 = nn.Sequential(
-//                nn.Conv2d(in_channels, out_channels, kernel_size = 3, stride = stride, padding = 1),
-//                nn.BatchNorm2d(out_channels),
-//                nn.ReLU())
-//        self.conv2 = nn.Sequential(
-//                nn.Conv2d(out_channels, out_channels, kernel_size = 3, stride = 1, padding = 1),
-//                nn.BatchNorm2d(out_channels))
-//        self.downsample = downsample
-//        self.relu = nn.ReLU()
-//        self.out_channels = out_channels
-//
-//    def forward(self, x):
-//        residual = x
-//        out = self.conv1(x)
-//        out = self.conv2(out)
-//        if self.downsample:
-//            residual = self.downsample(x)
-//        out += residual
-//                out = self.relu(out)
-//        return out
-
-struct ResidualBlock : torch::nn::Module {
[…]
