+++ examples/cnn/alexnet-cifar100.cpp
@@ -1,5 +1,5 @@
 #include "includes/base.h"
-#include "../../include/datasets/specific/cifar.h"
+#include "../../include/datasets/image-classification/cifar.h"
 #include "../../include/models/cnn/alexnet.h"
 #include "../../include/definitions/transforms.h"
 
