+++ temp/scvdatasettest.cpp
@@ -9,7 +9,7 @@
 #include "../include/definitions/transforms.h"
 
 #include "../include/datasets/specific/mnist.h"
-#include "../include/datasets/specific/cifar.h"
+#include "../include/datasets/image-classification/cifar.h"
 #include "../include/datasets/specific/imagenette.h"
 #include "../include/models/cnn/lenet5.h"
 #include "../include/definitions/transforms.h"
