+++ examples/cnn/resnet-cifar10.cpp
@@ -1,5 +1,5 @@
 #include "includes/base.h"
-#include "../../include/datasets/specific/cifar.h"
+#include "../../include/datasets/image-classification/cifar.h"
 #include "../../include/models/cnn/resnet.h"
 #include "../../include/definitions/transforms.h"
 
