+++ examples/cnn/vggnet16-cifar10.cpp
@@ -1,6 +1,6 @@
 #include "includes/base.h"
 
-#include "../../include/datasets/specific/cifar.h"
+#include "../../include/datasets/image-classification/cifar.h"
 #include "../../include/models/cnn/vggnet.h"
 #include "../../include/definitions/transforms.h"
 
