+++ examples/cnn/resnet-cifar100.cpp
@@ -1,6 +1,6 @@
 #include "includes/base.h"
 
-#include "../../include/datasets/specific/cifar.h"
+#include "../../include/datasets/image-classification/cifar.h"
 #include "../../include/models/cnn/resnet.h"
 #include "../../include/definitions/transforms.h"
 
