+++ examples/cnn/vggnet16-food101.cpp
@@ -1,5 +1,5 @@
 #include "includes/base.h"
-#include "../../include/datasets/specific/food.h"
+#include "../../include/datasets/image-classification/food.h"
 #include "../../include/models/cnn/vggnet.h"
 #include "../../include/definitions/transforms.h"
 
