+++ examples/cnn/lenet5-food101.cpp
@@ -1,5 +1,5 @@
 #include "includes/base.h"
-#include "../../include/datasets/specific/food.h"
+#include "../../include/datasets/image-classification/food.h"
 #include "../../include/models/cnn/lenet5.h"
 #include "../../include/definitions/transforms.h"
 
