+++ temp/dltest4.cpp
@@ -3,7 +3,7 @@
 #include <vector>
 #include <algorithm>
 #include <random>
-#include "../include/datasets/specific/mnist.h"
+#include "../include/datasets/image-classification/mnist.h"
 #include "../include/models/cnn/lenet5.h"
 #include "../include/definitions/transforms.h"
 #include <torch/torch.h>
