+++ examples/cnn/lenet5-fashionmnist.cpp
@@ -1,5 +1,5 @@
 #include "includes/base.h"
-#include "../../include/datasets/specific/mnist.h"
+#include "../../include/datasets/image-classification/mnist.h"
 #include "../../include/models/cnn/lenet5.h"
 #include "../../include/definitions/transforms.h"
 
