+++ examples/trainer/simple.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 #include <vector>
 #include <functional>
-#include "../../include/datasets/specific/mnist.h"
+#include "../../include/datasets/image-classification/mnist.h"
 #include "../../include/models/cnn/lenet5.h"
 #include "../../include/definitions/transforms.h"
 #include "../../include/data-loaders/data-loader.h"
