+++ examples/cnn/resnet-imagenette.cpp
@@ -1,5 +1,5 @@
 #include "includes/base.h"
-#include "../../include/datasets/specific/imagenette.h"
+#include "../../include/datasets/image-classification/imagenette.h"
 #include "../../include/models/cnn/resnet.h"
 #include "../../include/definitions/transforms.h"
 
