+++ include/datasets/audio-speech/common-voice.h
@@ -1,0 +1,8 @@
+//
+// Created by pc on 4/14/2025.
+//
+
+#ifndef XTORCH_COMMON_VOICE_H
+#define XTORCH_COMMON_VOICE_H
+
+#endif //XTORCH_COMMON_VOICE_H
