+++ include/datasets/text-classification/ag-news.h
@@ -1,0 +1,8 @@
+//
+// Created by pc on 4/14/2025.
+//
+
+#ifndef XTORCH_AG_NEWS_H
+#define XTORCH_AG_NEWS_H
+
+#endif //XTORCH_AG_NEWS_H
