### 2025-04-14 38ed54e
CMakeLists.txt updated and bugs fixed
