+++ include/datasets/audio-speech/yes-no.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/15/25.
+//
+
+#ifndef YES_NO_H
+#define YES_NO_H
+
+#endif //YES_NO_H
