+++ include/datasets/audio-speech/speech-commands.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/15/25.
+//
+
+#ifndef SPEECH_COMMANDS_H
+#define SPEECH_COMMANDS_H
+
+#endif //SPEECH_COMMANDS_H
