+++ include/datasets/audio-speech/gtzan.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/15/25.
+//
+
+#ifndef GTZAN_H
+#define GTZAN_H
+
+#endif //GTZAN_H
