+++ include/datasets/audio-speech/lj-speech.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/15/25.
+//
+
+#ifndef LJ_SPEECH_H
+#define LJ_SPEECH_H
+
+#endif //LJ_SPEECH_H
