+++ include/datasets/audio-speech/timit.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/15/25.
+//
+
+#ifndef TIMIT_H
+#define TIMIT_H
+
+#endif //TIMIT_H
