+++ include/datasets/text-classification/sogou-news.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/15/25.
+//
+
+#ifndef SOGOU_NEWS_H
+#define SOGOU_NEWS_H
+
+#endif //SOGOU_NEWS_H
