+++ include/datasets/text-classification/db-pedia.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/15/25.
+//
+
+#ifndef DB_PEDIA_H
+#define DB_PEDIA_H
+
+#endif //DB_PEDIA_H
