+++ include/datasets/text-classification/yahoo-answers.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/15/25.
+//
+
+#ifndef YAHOO_ANSWERS_H
+#define YAHOO_ANSWERS_H
+
+#endif //YAHOO_ANSWERS_H
