+++ include/datasets/text-classification/amazon-review.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/15/25.
+//
+
+#ifndef AMAZON_REVIEW_H
+#define AMAZON_REVIEW_H
+
+#endif //AMAZON_REVIEW_H
