+++ src/datasets/optical-flow​/flying-chairs.cpp
@@ -20,28 +20,4 @@ namespace xt::data::datasets {
         throw std::runtime_error("FlyingChairs: FlyingChairs not implemented");
     }
 
-
-
-
-
-
-    // ---------------------- FlyingThings3D ---------------------- //
-
-
-    FlyingThings3D::FlyingThings3D(const std::string &root): FlyingThings3D::FlyingThings3D(root, DataMode::TRAIN, false) {
-    }
-
-    FlyingThings3D::FlyingThings3D(const std::string &root, DataMode mode): FlyingThings3D::FlyingThings3D(root, mode, false) {
-    }
-
-    FlyingThings3D::FlyingThings3D(const std::string &root, DataMode mode, bool download) : BaseDataset(root, mode, download) {
-        throw std::runtime_error("FlyingThings3D: FlyingThings3D not implemented");
-    }
-
-
-    FlyingThings3D::FlyingThings3D(const std::string &root, DataMode mode, bool download,
-                           TransformType transforms) : BaseDataset(root, mode, download, transforms) {
-        throw std::runtime_error("FlyingThings3D: FlyingThings3D not implemented");
-    }
-
 }
