+++ src/models/cnn/mobilenet.cpp
@@ -23,130 +23,130 @@ namespace xt::models {
 
     // --------------------------------------------------------------------
 
-    SqueezeExcite::SqueezeExcite(int input_channels, int squeeze) {
-        this->SE = torch::nn::Sequential(
-            torch::nn::AdaptiveAvgPool2d(torch::nn::AdaptiveAvgPool2dOptions(1)),
-            torch::nn::Conv2d(
-                torch::nn::Conv2dOptions(input_channels, input_channels / squeeze, 1).stride(1).bias(false)),
-            torch::nn::BatchNorm2d(input_channels / squeeze),
-            torch::nn::ReLU(torch::nn::ReLUOptions(true)),
-            torch::nn::Conv2d(
-                torch::nn::Conv2dOptions(input_channels / squeeze, input_channels, 1).stride(1).bias(false)),
-            torch::nn::BatchNorm2d(input_channels),
-            HSigmoid::HSigmoid()
-        );
-    }
+     SqueezeExcite::SqueezeExcite(int input_channels, int squeeze) {
+         this->SE = torch::nn::Sequential(
+             torch::nn::AdaptiveAvgPool2d(torch::nn::AdaptiveAvgPool2dOptions(1)),
+             torch::nn::Conv2d(
+                 torch::nn::Conv2dOptions(input_channels, input_channels / squeeze, 1).stride(1).bias(false)),
+             torch::nn::BatchNorm2d(input_channels / squeeze),
+             torch::nn::ReLU(torch::nn::ReLUOptions(true)),
+             torch::nn::Conv2d(
+                 torch::nn::Conv2dOptions(input_channels / squeeze, input_channels, 1).stride(1).bias(false)),
+             torch::nn::BatchNorm2d(input_channels),
+             HSigmoid()
+         );
+     }
 
-    torch::Tensor SqueezeExcite::forward(torch::Tensor x) {
-        x = this->SE->forward(x);
-        return x;
-    }
-
-    // --------------------------------------------------------------------
-
-    Bottleneck::Bottleneck(int input_channels, int kernel, int stride, int expansion, int output_channels,
-                           torch::nn::Module activation, bool se) {
-        this->bottleneck = torch::nn::Sequential(
-            torch::nn::Conv2d(
-                torch::nn::Conv2dOptions(input_channels, expansion, 1).stride(1).bias(false)),
-            torch::nn::BatchNorm2d(expansion),
-            activation,
-
-
-            torch::nn::Conv2d(
-                torch::nn::Conv2dOptions(expansion, expansion, kernel).stride(stride).padding(kernel / 2).
-                groups(expansion).bias(false)),
-            torch::nn::BatchNorm2d(expansion),
-            activation,
-
-            SqueezeExcite(expansion),
-
-
-            torch::nn::Conv2d(
-                torch::nn::Conv2dOptions(expansion, output_channels, 1).stride(1).bias(false)),
-            torch::nn::BatchNorm2d(expansion),
-            activation
-        );
-
-
-        if (input_channels == output_channels && stride == 1) {
-            this->downsample = torch::nn::Sequential();
-        } else {
-            this->downsample = torch::nn::Sequential(
-                torch::nn::Conv2d(
-                    torch::nn::Conv2dOptions(input_channels, output_channels, 1).stride(stride).bias(false)),
-                torch::nn::BatchNorm2d(expansion)
-            );
-        }
-    }
-
-    torch::Tensor Bottleneck::forward(torch::Tensor x) {
-        torch::Tensor residual = x;
[…]
