### 2025-04-17 5b99b61
some bugs raised while grouping datasets . some of the fixed #2
