### 2025-04-17 c78fe0f
some bugs raised while grouping datasets . some of the fixed #3
