### 2025-04-17 18d31d3
some bugs raised while grouping datasets . some of the fixed #3
