+++ src/datasets/image-classification/places.cpp
@@ -1,4 +1,4 @@
-#include "../../../include/datasets/specific/places.h"
+#include "../../../include/datasets/image-classification/places.h"
 
 namespace xt::data::datasets {
 
