### 2025-04-17 3438e15
some bugs raised while grouping datasets . some of the fixed #4
