### 2025-04-17 a3e7d39
some bugs raised while grouping datasets . some of the fixed #5
