+++ src/datasets/image-classification/sun.cpp
@@ -1,4 +1,4 @@
-#include "../../../include/datasets/specific/sun.h"
+#include "../../../include/datasets/image-classification/sun.h"
 
 namespace xt::data::datasets {
 
