### 2025-04-17 6c802ca
some bugs raised while grouping datasets . some of the fixed #6
