### 2025-04-17 19b0cfb
some bugs raised while grouping datasets . some of the fixed #7
