### 2025-04-17 6bcbc26
some bugs raised while grouping datasets . some of the fixed #7
