### 2025-04-17 7330a66
some bugs raised while grouping datasets . some of the fixed #8
