+++ src/datasets/stereo-matching-depth-estimation/cre-stereo.cpp
@@ -1,4 +1,4 @@
-#include "../../../include/datasets/specific/cre-stereo.h"
+#include "../../../include/datasets/stereo-matching-depth-estimation/cre-stereo.h"
 
 namespace xt::data::datasets {
     // ---------------------- CarlaStereo ---------------------- //
