### 2025-04-17 eabfb1a
some bugs raised while grouping datasets . some of the fixed #9
