+++ src/datasets/stereo-matching-depth-estimation/eth-3d-stereo.cpp
@@ -1,4 +1,4 @@
-#include "../../../include/datasets/specific/eth-3d-stereo.h"
+#include "../../../include/datasets/stereo-matching-depth-estimation/eth-3d-stereo.h"
 
 namespace xt::data::datasets {
 
