### 2025-04-17 8669a06
some bugs raised while grouping datasets . some of the fixed #10
