### 2025-04-17 a6ff462
some bugs raised while grouping datasets . some of the fixed #10
