+++ src/datasets/stereo-matching-depth-estimation/scene-flow-stereo.cpp
@@ -1,4 +1,4 @@
-#include "../../../include/datasets/specific/scene-flow-stereo.h"
+#include "../../../include/datasets/stereo-matching-depth-estimation/scene-flow-stereo.h"
 
 namespace xt::data::datasets {
 
