+++ include/xtorch.h
@@ -1,11 +1,2 @@
-//
-// Created by kami on 2/8/25.
-//
+#pragma once
 
-#ifndef DCGAN_BASE_H
-#define DCGAN_BASE_H
-
-#include "utils/base.h"
-#include "datasets/base.h"
-
-#endif //DCGAN_BASE_H
