+++ include/types/types.h
@@ -1,0 +1,5 @@
+#pragma once
+
+
+#include "arguments.h"
+#include "enums.h"
