+++ include/transforms/transforms.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "compose.h"
+#include "convert.h"
+#include "crop.h"
+#include "cutout.h"
+#include "flip.h"
+#include "gaussian.h"
+#include "grayscale.h"
+#include "jitter.h"
+#include "lambda.h"
+#include "normalize.h"
+#include "pad.h"
+#include "resize.h"
+#include "rotation.h"
+
+
+
