+++ include/xtorch.h
@@ -2,3 +2,4 @@
 
 #include "utils/utils.h"
 #include "types/types.h"
+#include "transforms/transforms.h"
