+++ include/xtorch.h
@@ -3,3 +3,5 @@
 #include "utils/utils.h"
 #include "types/types.h"
 #include "transforms/transforms.h"
+#include "trainers/trainers.h"
+
