+++ include/models/models.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#include "base.h"
+#include "cnn/cnn.h"
