+++ include/xtorch.h
@@ -4,4 +4,5 @@
 #include "types/types.h"
 #include "transforms/transforms.h"
 #include "trainers/trainers.h"
+#include "models/models.h"
 
