+++ include/models/cnn/cnn.h
@@ -1,0 +1,9 @@
+#pragma once
+
+
+#include "alexnet.h"
+#include "lenet5.h"
+#include "mobilenet.h"
+#include "resnet.h"
+#include "unet.h"
+#include "vggnet.h"
