+++ include/media/opencv/opencv.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#include "images.h"
+#include "videos.h"
