+++ include/xtorch.h
@@ -5,4 +5,5 @@
 #include "transforms/transforms.h"
 #include "trainers/trainers.h"
 #include "models/models.h"
+#include "media/media.h"
 
