+++ include/media/media.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "image/image.h"
+#include "opencv/opencv.h"
+#include "video/video.h"
