+++ include/media/image/image.h
@@ -1,0 +1,2 @@
+#pragma once
+#include "processing.h"
