+++ include/media/video/video.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "processing.h"
