+++ include/data-loaders/data-loaders.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "data-loader.h"
