+++ include/datasets/general/general.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/18/25.
+//
+
+#ifndef GENERAL_H
+#define GENERAL_H
+
+#endif //GENERAL_H
