### 2025-04-17 c6db286
any foler in include folder will gets a header file with same name of folder to keep all include files inside it #7
