+++ include/datasets/face-recognition/face-recognition.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/18/25.
+//
+
+#ifndef FACE_RECOGNITION_H
+#define FACE_RECOGNITION_H
+
+#endif //FACE_RECOGNITION_H
