+++ include/datasets/image-classification/image-classification.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/18/25.
+//
+
+#ifndef IMAGE_CLASSIFICATION_H
+#define IMAGE_CLASSIFICATION_H
+
+#endif //IMAGE_CLASSIFICATION_H
