+++ include/datasets/object-detection-and-segmentation/object-detection-and-segmentation.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/18/25.
+//
+
+#ifndef OBJECT_DETECTION_AND_SEGMENTATION_H
+#define OBJECT_DETECTION_AND_SEGMENTATION_H
+
+#endif //OBJECT_DETECTION_AND_SEGMENTATION_H
