+++ include/datasets/sequence-tagging/sequence-tagging.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/18/25.
+//
+
+#ifndef SEQUENCE_TAGGING_H
+#define SEQUENCE_TAGGING_H
+
+#endif //SEQUENCE_TAGGING_H
