+++ include/datasets/sentiment-analysis/sentiment-analysis.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/18/25.
+//
+
+#ifndef SENTIMENT_ANALYSIS_H
+#define SENTIMENT_ANALYSIS_H
+
+#endif //SENTIMENT_ANALYSIS_H
