+++ include/datasets/stereo-matching-depth-estimation/stereo-matching-depth-estimation.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/18/25.
+//
+
+#ifndef STEREO_MATCHING_DEPTH_ESTIMATION_H
+#define STEREO_MATCHING_DEPTH_ESTIMATION_H
+
+#endif //STEREO_MATCHING_DEPTH_ESTIMATION_H
