+++ include/datasets/video/video.h
@@ -1,0 +1,8 @@
+//
+// Created by kami on 4/18/25.
+//
+
+#ifndef VIDEO_H
+#define VIDEO_H
+
+#endif //VIDEO_H
