+++ include/datasets/datasets.h
@@ -1,3 +1,21 @@
 #pragma once
 
 #include "audio-speech/audio-speech.h"
+#include "base/base.h"
+#include "definitions/definitions.h"
+#include "face-recognition/face-recognition.h"
+#include "general/general.h"
+#include "image-classification/image-classification.h"
+#include "language-modeling/language-modeling.h"
+#include "machine-translation/machine-translation.h"
+#include "natural-language-inference/natural-language-inference.h"
+#include "object-detection-and-segmentation/object-detection-and-segmentation.h"
+#include "optical-flow/optical-flow.h"
+#include "patch-matching-descriptor-learning/patch-matching-descriptor-learning.h"
+#include "scene-flow/scene-flow.h"
+#include "sentiment-analysis/sentiment-analysis.h"
+#include "sequence-tagging/sequence-tagging.h"
+#include "specific/specific.h"
+#include "stereo-matching-depth-estimation/stereo-matching-depth-estimation.h"
+#include "text-classification/text-classification.h"
+#include "video/video.h"
