+++ include/xtorch.h
@@ -1,10 +1,15 @@
 #pragma once
 
-#include "utils/utils.h"
-#include "types/types.h"
-#include "transforms/transforms.h"
-#include "trainers/trainers.h"
-#include "models/models.h"
-#include "media/media.h"
 #include "data-loaders/data-loaders.h"
+#include "datasets/datasets.h"
+#include "media/media.h"
+#include "models/models.h"
+#include "trainers/trainers.h"
+#include "transforms/transforms.h"
+#include "types/types.h"
+#include "utils/utils.h"
+
+
+
+
 
