### 2025-04-17 8822508
audio-speech.h created
