### 2025-04-17 348a7ac
face-recognition.h created
