+++ include/datasets/definitions/definitions.h
@@ -1,8 +1,2 @@
-//
-// Created by kami on 4/18/25.
-//
-
-#ifndef DEFINITIONS_H
-#define DEFINITIONS_H
-
-#endif //DEFINITIONS_H
+#pragma once
+#include "functions.h"
