### 2025-04-17 434f6fc
general.h created
