+++ include/datasets/image-classification/image-classification.h
@@ -1,8 +1,29 @@
-//
-// Created by kami on 4/18/25.
-//
+#pragma once
 
-#ifndef IMAGE_CLASSIFICATION_H
-#define IMAGE_CLASSIFICATION_H
-
-#endif //IMAGE_CLASSIFICATION_H
+#include "caltech.h"
+#include "celeba.h"
+#include "cifar.h"
+#include "country.h"
+#include "dtd.h"
+#include "euro-sat.h"
+#include "fake-data.h"
+#include "fer.h"
+#include "fgvc-aircraft.h"
+#include "flickr.h"
+#include "flowers.h"
+#include "food.h"
+#include "gtsrb.h"
+#include "imagenet.h"
+#include "imagenette.h"
+#include "inaturalist.h"
+#include "mnist.h"
+#include "omniglot.h"
+#include "pcam.h"
+#include "places.h"
+#include "rendered-sst2.h"
+#include "semeion.h"
+#include "stanford-cars.h"
+#include "stl.h"
+#include "sun.h"
+#include "svhn.h"
+#include "usps.h"
