### 2025-04-17 3c5fa95
machine-translation.h created
