### 2025-04-17 0be4b3b
object-detection-and-segmentation.h created
