### 2025-04-17 01bd1fa
sentiment-analysis.h created
