+++ include/datasets/sentiment-analysis/sentiment-analysis.h
@@ -1,8 +1,4 @@
-//
-// Created by kami on 4/18/25.
-//
+#pragma once
 
-#ifndef SENTIMENT_ANALYSIS_H
-#define SENTIMENT_ANALYSIS_H
-
-#endif //SENTIMENT_ANALYSIS_H
+#include "imdb.h"
+#include "sst.h"
