+++ include/datasets/sequence-tagging/sequence-tagging.h
@@ -1,8 +1,4 @@
-//
-// Created by kami on 4/18/25.
-//
+#pragma once
 
-#ifndef SEQUENCE_TAGGING_H
-#define SEQUENCE_TAGGING_H
-
-#endif //SEQUENCE_TAGGING_H
+#include "Co-ll-2000-chunking.h"
+#include "udpos.h"
