### 2025-04-17 9c106fa
specific.h created
