+++ include/datasets/specific/specific.h
@@ -1,8 +1,6 @@
-//
-// Created by kami on 4/18/25.
-//
-
-#ifndef SPECIFIC_H
-#define SPECIFIC_H
-
-#endif //SPECIFIC_H
+#pragma once
+#include "kitti.h"
+#include "lsun.h"
+#include "sbu.h"
+#include "sintel.h"
+#include "voc.h"
