+++ include/datasets/text-classification/text-classification.h
@@ -1,8 +1,8 @@
-//
-// Created by kami on 4/18/25.
-//
+#pragma once
 
-#ifndef TEXT_CLASSIFICATION_H
-#define TEXT_CLASSIFICATION_H
-
-#endif //TEXT_CLASSIFICATION_H
+#include "ag-news.h"
+#include "amazon-review.h"
+#include "db-pedia.h"
+#include "sogou-news.h"
+#include "yahoo-answers.h"
+#include "yelp-review.h"
