+++ include/data-loaders/data-loader.h
@@ -66,5 +66,5 @@ private:
 
 
 }
-// #include "../../src/data-loaders/data-loader.tpp"
+#include "data-loader.tpp"
 
