### 2025-04-18 a0056cc
sudo make uninstall added to library
