+++ cmake_uninstall.cmake.in
@@ -1,0 +1,23 @@
+# cmake_uninstall.cmake.in
+if(NOT EXISTS "@CMAKE_CURRENT_BINARY_DIR@/install_manifest.txt")
+  message(FATAL_ERROR "Cannot find install manifest: @CMAKE_CURRENT_BINARY_DIR@/install_manifest.txt")
+endif()
+
+file(READ "@CMAKE_CURRENT_BINARY_DIR@/install_manifest.txt" files)
+string(REGEX REPLACE "\n" ";" files "${files}")
+
+foreach(file ${files})
+  if(EXISTS "${file}")
+    message(STATUS "Removing: ${file}")
+    exec_program(
+      "@CMAKE_COMMAND@" ARGS "-E remove \"${file}\""
+      OUTPUT_VARIABLE rm_out
+      RETURN_VALUE rm_retval
+    )
+    if(NOT "${rm_retval}" STREQUAL 0)
+      message(FATAL_ERROR "Failed to remove: ${file}")
+    endif()
+  else()
+    message(STATUS "File does not exist: ${file}")
+  endif()
+endforeach()
