+++ examples/CMakeLists.txt
@@ -357,7 +357,7 @@ set(SOURCE_FILES
         ../src/datasets/image-classification/mnist.cpp
         ../src/models/cnn/lenet5.cpp
         ../src/trainers/trainer.cpp
-        ../src/data-loaders/data-loader.tpp
+#        ../src/data-loaders/data-loader.tpp
 
 )
 add_executable(trainer-simple ${BASE_SOURCE_FILES} ${SOURCE_FILES} trainer/simple.cpp)
