### 2025-04-18 6d8fba5
sudo make uninstall added to library
