### 2025-04-18 73cc2bb
sudo make uninstall added to library
