+++ include/datasets/general/tensor-dataset.h
@@ -22,18 +22,6 @@ namespace xt::data::datasets {
     };
 
 
-    class StackedCSVDataset : public BaseDataset {
-    public :
-        StackedCSVDataset(const std::string &folder_path);
-        StackedCSVDataset(const std::string &folder_path,DataMode mode);
-        StackedCSVDataset(const std::string &folder_path,DataMode mode, bool load_sub_folders);
-        StackedCSVDataset(const std::string &folder_path,DataMode mode, bool load_sub_folders , vector<std::function<torch::Tensor(torch::Tensor)>> transforms);
-
-    private:
-        vector<string> labels_name;
-        bool load_sub_folders = false;
-        void load_data();
-    };
 
 
 
