+++ include/datasets/general/time-series-dataset.h
@@ -22,12 +22,12 @@ namespace xt::data::datasets {
     };
 
 
-    class StackedCSVDataset : public BaseDataset {
+    class StackedTimeSeriesDataset : public BaseDataset {
     public :
-        StackedCSVDataset(const std::string &folder_path);
-        StackedCSVDataset(const std::string &folder_path,DataMode mode);
-        StackedCSVDataset(const std::string &folder_path,DataMode mode, bool load_sub_folders);
-        StackedCSVDataset(const std::string &folder_path,DataMode mode, bool load_sub_folders , vector<std::function<torch::Tensor(torch::Tensor)>> transforms);
+        StackedTimeSeriesDataset(const std::string &folder_path);
+        StackedTimeSeriesDataset(const std::string &folder_path,DataMode mode);
+        StackedTimeSeriesDataset(const std::string &folder_path,DataMode mode, bool load_sub_folders);
+        StackedTimeSeriesDataset(const std::string &folder_path,DataMode mode, bool load_sub_folders , vector<std::function<torch::Tensor(torch::Tensor)>> transforms);
 
     private:
         vector<string> labels_name;
