+++ include/datasets/general/text-dataset.h
@@ -22,12 +22,12 @@ namespace xt::data::datasets {
     };
 
 
-    class StackedCSVDataset : public BaseDataset {
+    class StackedTextDataset : public BaseDataset {
     public :
-        StackedCSVDataset(const std::string &folder_path);
-        StackedCSVDataset(const std::string &folder_path,DataMode mode);
-        StackedCSVDataset(const std::string &folder_path,DataMode mode, bool load_sub_folders);
-        StackedCSVDataset(const std::string &folder_path,DataMode mode, bool load_sub_folders , vector<std::function<torch::Tensor(torch::Tensor)>> transforms);
+        StackedTextDataset(const std::string &folder_path);
+        StackedTextDataset(const std::string &folder_path,DataMode mode);
+        StackedTextDataset(const std::string &folder_path,DataMode mode, bool load_sub_folders);
+        StackedTextDataset(const std::string &folder_path,DataMode mode, bool load_sub_folders , vector<std::function<torch::Tensor(torch::Tensor)>> transforms);
 
     private:
         vector<string> labels_name;
