+++ include/datasets/general/audio-dataset.h
@@ -50,5 +50,23 @@ namespace xt::data::datasets {
         std::vector<int> labels;
     };
 
+    class StackedAudioDataset : public BaseDataset {
+        public :
+            StackedAudioDataset(const std::string &folder_path);
+
+        StackedAudioDataset(const std::string &folder_path, DataMode mode);
+
+        StackedAudioDataset(const std::string &folder_path, DataMode mode, bool load_sub_folders);
+
+        StackedAudioDataset(const std::string &folder_path, DataMode mode, bool load_sub_folders,
+                          vector<std::function<torch::Tensor(torch::Tensor)> > transforms);
+
+    private:
+        vector<string> labels_name;
+        bool load_sub_folders = false;
+
+        void load_data();
+    };
+
 
 }
