+++ include/temp/test-dataset.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include <torch/torch.h>
+
+namespace xt::temp {
+
+class TestDataset {
+
+  public:
+    TestDataset();
+    ~TestDataset();
+    torch::Tensor forward(torch::Tensor input);
+};
+
+}
+
