+++ src/temp/test-dataset.cpp
@@ -1,0 +1,12 @@
+#include "../../include/temp/test-dataset.h"
+
+namespace xt::temp {
+
+    TestDataset::TestDataset(){}
+    TestDataset::~TestDataset(){}
+
+    torch::Tensor forward(torch::Tensor input){
+        return input;
+    }
+
+}
