+++ CMakeLists-final.txt
@@ -1,0 +1,240 @@
+# -----------------------------------------------------------------------------
+# Project Setup for xTorch
+# -----------------------------------------------------------------------------
+# This CMake configuration file sets up the build environment for the xTorch
+# project, a C++ library. It specifies required dependencies, includes, and
+# configuration for building the project. It also handles library and
+# executable installation.
+#
+# Version: 0.2.0
+# CMake Minimum Version: 3.28
+#
+# -----------------------------------------------------------------------------
+# CMake Configuration
+# -----------------------------------------------------------------------------
+cmake_minimum_required(VERSION 3.28 FATAL_ERROR)
+
+# Define the project name and version
+project(xTorch VERSION 0.2.0 LANGUAGES CXX)
+
+# -----------------------------------------------------------------------------
+# Source Files
+# -----------------------------------------------------------------------------
+# Lists all the source files for various components of the project:
+#   - Utility files
+#   - Media files
+#   - Dataset files
+#   - Definition files
+#   - Model files
+#   - Dataloader files
+#   - Transform files
+file(GLOB_RECURSE UTILITY_FILES CONFIGURE_DEPENDS src/utils/*.cpp)
+file(GLOB_RECURSE MEDIA_FILES CONFIGURE_DEPENDS src/media/*.cpp)
+file(GLOB_RECURSE DATASET_FILES CONFIGURE_DEPENDS src/datasets/*.cpp)
+file(GLOB_RECURSE DEFINITION_FILES CONFIGURE_DEPENDS src/definitions/*.cpp)
+file(GLOB_RECURSE MODEL_FILES CONFIGURE_DEPENDS src/models/*.cpp)
+file(GLOB_RECURSE MODEL_CNN_FILES CONFIGURE_DEPENDS src/models/cnn/*.cpp)
+file(GLOB_RECURSE DATALOADER_FILES CONFIGURE_DEPENDS src/data-loaders/*.tpp)
+file(GLOB_RECURSE TRANSFORM_FILES CONFIGURE_DEPENDS src/transforms/*.cpp)
+
+
+
+file(GLOB_RECURSE TEMP_FILES CONFIGURE_DEPENDS src/temp/*.cpp)
+
+
+message(${TEMP_FILES})
+
+# Combine all source files into one variable for easy access
+set(LIBRARY_SOURCE_FILES ${UTILITY_FILES} ${MEDIA_FILES} ${DATASET_FILIES} ${DEFINITION_FILES} ${MODEL_FILES}  ${DATALOADER_FILES} ${TRANSFORM_FILES} ${DATASET_FILES} ${TEMP_FILES} )
+
+# -----------------------------------------------------------------------------
+# Library Setup
+# -----------------------------------------------------------------------------
+# Create the shared library from the source files
+add_library(${CMAKE_PROJECT_NAME} SHARED ${LIBRARY_SOURCE_FILES})
+
+# -----------------------------------------------------------------------------
+# External Dependencies
+# -----------------------------------------------------------------------------
+# Find required libraries and header files for dependencies such as libtar,
+# libtorch, OpenCV, Zlib, OpenSSL, LibLZMA, etc.
+find_library(LIBTAR_LIBRARY tar REQUIRED)
+find_path(LIBTAR_INCLUDE_DIR tar.h REQUIRED)
+
+# Ensure that libtar is found, else terminate the build
[…]
