### 2025-04-18 017721b
all bugs fixed and simple test updated
