+++ include/datasets/audio-speech/cmu-arctic.h
@@ -1,8 +1,19 @@
-//
-// Created by kami on 4/15/25.
-//
+#pragma once
+#include "../base/base.h"
+#include "../../headers/datasets.h"
 
-#ifndef CMU_ARCTIC_H
-#define CMU_ARCTIC_H
 
-#endif //CMU_ARCTIC_H
+namespace xt::data::datasets {
+    class CMUArctic : BaseDataset {
+    public :
+        explicit CMUArctic(const std::string &root);
+        CMUArctic(const std::string &root, DataMode mode);
+        CMUArctic(const std::string &root, DataMode mode , bool download);
+        CMUArctic(const std::string &root, DataMode mode , bool download, TransformType transforms);
+
+    private :
+        void load_data();
+
+        void check_resources();
+    };
+}
