### 2025-04-18 5972872
all bugs fixed and simple test updated
