+++ CMakeLists.txt
@@ -105,3 +105,12 @@ if(NOT TARGET uninstall)
     add_custom_target(uninstall
             COMMAND ${CMAKE_COMMAND} -P ${CMAKE_CURRENT_BINARY_DIR}/cmake_uninstall.cmake)
 endif()
+
+
+
+# -----------------------------------------------------------------------------
+# Subdirectories
+# -----------------------------------------------------------------------------
+# Add subdirectories for example and temporary code
+add_subdirectory(examples)
+#add_subdirectory(temp)
