+++ include/datasets/audio-speech/esc.h
@@ -1,8 +1,23 @@
-//
-// Created by kami on 4/15/25.
-//
+#pragma once
 
-#ifndef ESC_H
-#define ESC_H
+#include "../base/base.h"
+#include "../../headers/datasets.h"
 
-#endif //ESC_H
+
+namespace xt::data::datasets {
+    class ESC : BaseDataset {
+    public :
+        explicit ESC(const std::string &root);
+
+        ESC(const std::string &root, DataMode mode);
+
+        ESC(const std::string &root, DataMode mode, bool download);
+
+        ESC(const std::string &root, DataMode mode, bool download, TransformType transforms);
+
+    private :
+        void load_data();
+
+        void check_resources();
+    };
+}
