+++ src/datasets/audio-speech/cmu-arctic.cpp
@@ -9,13 +9,13 @@ namespace xt::data::datasets {
     }
 
     CMUArctic::CMUArctic(const std::string &root, DataMode mode, bool download) : BaseDataset(root, mode, download) {
-        throw std::runtime_error("GTSRB: GTSRB not implemented");
+        throw std::runtime_error("CMUArctic: CMUArctic not implemented");
     }
 
 
     CMUArctic::CMUArctic(const std::string &root, DataMode mode, bool download,
                            TransformType transforms) : BaseDataset(root, mode, download, transforms) {
-        throw std::runtime_error("GTSRB: GTSRB not implemented");
+        throw std::runtime_error("CMUArctic: CMUArctic not implemented");
     }
 
 
