+++ src/datasets/audio-speech/common-voice.cpp
@@ -13,13 +13,13 @@ namespace xt::data::datasets {
     }
 
     CommonVoice::CommonVoice(const std::string &root, DataMode mode, bool download) : BaseDataset(root, mode, download) {
-        throw std::runtime_error("GTSRB: GTSRB not implemented");
+        throw std::runtime_error("CommonVoice: CommonVoice not implemented");
     }
 
 
     CMUArctic::CMUArctic(const std::string &root, DataMode mode, bool download,
                          TransformType transforms) : BaseDataset(root, mode, download, transforms) {
-        throw std::runtime_error("GTSRB: GTSRB not implemented");
+        throw std::runtime_error("CommonVoice: CommonVoice not implemented");
     }
 
 
