+++ src/datasets/audio-speech/crema-d.cpp
@@ -9,13 +9,13 @@ namespace xt::data::datasets {
     }
 
     CremaD::CremaD(const std::string &root, DataMode mode, bool download) : BaseDataset(root, mode, download) {
-        throw std::runtime_error("GTSRB: GTSRB not implemented");
+        throw std::runtime_error("CremaD: CremaD not implemented");
     }
 
 
     CremaD::CremaD(const std::string &root, DataMode mode, bool download,
                          TransformType transforms) : BaseDataset(root, mode, download, transforms) {
-        throw std::runtime_error("GTSRB: GTSRB not implemented");
+        throw std::runtime_error("CremaD: CremaD not implemented");
     }
 
 
