+++ include/datasets/audio-speech/gtzan.h
@@ -1,8 +1,23 @@
-//
-// Created by kami on 4/15/25.
-//
+#pragma once
 
-#ifndef GTZAN_H
-#define GTZAN_H
+#include "../base/base.h"
+#include "../../headers/datasets.h"
 
-#endif //GTZAN_H
+
+namespace xt::data::datasets {
+    class GTZAN : BaseDataset {
+    public :
+        explicit GTZAN(const std::string &root);
+
+        GTZAN(const std::string &root, DataMode mode);
+
+        GTZAN(const std::string &root, DataMode mode, bool download);
+
+        GTZAN(const std::string &root, DataMode mode, bool download, TransformType transforms);
+
+    private :
+        void load_data();
+
+        void check_resources();
+    };
+}
