+++ src/datasets/audio-speech/gtzan.cpp
@@ -1,3 +1,22 @@
-//
-// Created by kami on 4/15/25.
-//
+#include "../../../include/datasets/audio-speech/gtzan.h"
+
+namespace xt::data::datasets {
+
+    GTZAN::GTZAN(const std::string &root): CMUArctic::CMUArctic(root, DataMode::TRAIN, false) {
+    }
+
+    GTZAN::GTZAN(const std::string &root, DataMode mode): CMUArctic::CMUArctic(root, mode, false) {
+    }
+
+    GTZAN::GTZAN(const std::string &root, DataMode mode, bool download) : BaseDataset(root, mode, download) {
+        throw std::runtime_error("GTZAN: GTZAN not implemented");
+    }
+
+
+    GTZAN::GTZAN(const std::string &root, DataMode mode, bool download,
+                         TransformType transforms) : BaseDataset(root, mode, download, transforms) {
+        throw std::runtime_error("GTZAN: GTZAN not implemented");
+    }
+
+
+}
