+++ src/datasets/audio-speech/librispeech.cpp
@@ -1,8 +1,22 @@
-//
-// Created by pc on 4/14/2025.
-//
+#include "../../../include/datasets/audio-speech/librispeech.h"
 
-#ifndef XTORCH_LIBRISPEECH_H
-#define XTORCH_LIBRISPEECH_H
+namespace xt::data::datasets {
 
-#endif //XTORCH_LIBRISPEECH_H
+    LibriSpeech::LibriSpeech(const std::string &root): CMUArctic::CMUArctic(root, DataMode::TRAIN, false) {
+    }
+
+    LibriSpeech::LibriSpeech(const std::string &root, DataMode mode): CMUArctic::CMUArctic(root, mode, false) {
+    }
+
+    LibriSpeech::LibriSpeech(const std::string &root, DataMode mode, bool download) : BaseDataset(root, mode, download) {
+        throw std::runtime_error("LibriSpeech: LibriSpeech not implemented");
+    }
+
+
+    LibriSpeech::LibriSpeech(const std::string &root, DataMode mode, bool download,
+                         TransformType transforms) : BaseDataset(root, mode, download, transforms) {
+        throw std::runtime_error("LibriSpeech: GTSRB not implemented");
+    }
+
+
+}
