+++ src/datasets/audio-speech/timit.cpp
@@ -1,3 +1,23 @@
-//
-// Created by kami on 4/15/25.
-//
+#pragma once
+
+#include "../base/base.h"
+#include "../../headers/datasets.h"
+
+
+namespace xt::data::datasets {
+    class TIMIT : BaseDataset {
+    public :
+        explicit TIMIT(const std::string &root);
+
+        TIMIT(const std::string &root, DataMode mode);
+
+        TIMIT(const std::string &root, DataMode mode, bool download);
+
+        TIMIT(const std::string &root, DataMode mode, bool download, TransformType transforms);
+
+    private :
+        void load_data();
+
+        void check_resources();
+    };
+}
