+++ src/datasets/audio-speech/timit.cpp
@@ -1,23 +1,0 @@
-#pragma once
-
-#include "../base/base.h"
-#include "../../headers/datasets.h"
-
-
-namespace xt::data::datasets {
-    class TIMIT : BaseDataset {
-    public :
-        explicit TIMIT(const std::string &root);
-
-        TIMIT(const std::string &root, DataMode mode);
-
-        TIMIT(const std::string &root, DataMode mode, bool download);
-
-        TIMIT(const std::string &root, DataMode mode, bool download, TransformType transforms);
-
-    private :
-        void load_data();
-
-        void check_resources();
-    };
-}
