+++ src/datasets/audio-speech/vctk.cpp
@@ -1,3 +1,22 @@
-//
-// Created by kami on 4/15/25.
-//
+#include "../../../include/datasets/audio-speech/vctk.h"
+
+namespace xt::data::datasets {
+
+    CMUArctic::CMUArctic(const std::string &root): CMUArctic::CMUArctic(root, DataMode::TRAIN, false) {
+    }
+
+    CMUArctic::CMUArctic(const std::string &root, DataMode mode): CMUArctic::CMUArctic(root, mode, false) {
+    }
+
+    CMUArctic::CMUArctic(const std::string &root, DataMode mode, bool download) : BaseDataset(root, mode, download) {
+        throw std::runtime_error("CMUArctic: CMUArctic not implemented");
+    }
+
+
+    CMUArctic::CMUArctic(const std::string &root, DataMode mode, bool download,
+                           TransformType transforms) : BaseDataset(root, mode, download, transforms) {
+        throw std::runtime_error("CMUArctic: CMUArctic not implemented");
+    }
+
+
+}
