+++ src/datasets/audio-speech/vctk.cpp
@@ -2,20 +2,20 @@
 
 namespace xt::data::datasets {
 
-    CMUArctic::CMUArctic(const std::string &root): CMUArctic::CMUArctic(root, DataMode::TRAIN, false) {
+    VCTK::VCTK(const std::string &root): VCTK::VCTK(root, DataMode::TRAIN, false) {
     }
 
-    CMUArctic::CMUArctic(const std::string &root, DataMode mode): CMUArctic::CMUArctic(root, mode, false) {
+    VCTK::VCTK(const std::string &root, DataMode mode): VCTK::VCTK(root, mode, false) {
     }
 
-    CMUArctic::CMUArctic(const std::string &root, DataMode mode, bool download) : BaseDataset(root, mode, download) {
-        throw std::runtime_error("CMUArctic: CMUArctic not implemented");
+    VCTK::VCTK(const std::string &root, DataMode mode, bool download) : BaseDataset(root, mode, download) {
+        throw std::runtime_error("VCTK: VCTK not implemented");
     }
 
 
-    CMUArctic::CMUArctic(const std::string &root, DataMode mode, bool download,
+    VCTK::VCTK(const std::string &root, DataMode mode, bool download,
                            TransformType transforms) : BaseDataset(root, mode, download, transforms) {
-        throw std::runtime_error("CMUArctic: CMUArctic not implemented");
+        throw std::runtime_error("VCTK: VCTK not implemented");
     }
 
 
