+++ include/datasets/audio-speech/vox-celeb.h
@@ -1,8 +1,19 @@
-//
-// Created by kami on 4/15/25.
-//
+#pragma once
+#include "../base/base.h"
+#include "../../headers/datasets.h"
 
-#ifndef VOX_CELEB_H
-#define VOX_CELEB_H
 
-#endif //VOX_CELEB_H
+namespace xt::data::datasets {
+    class VoxCeleb : BaseDataset {
+        public :
+            explicit VoxCeleb(const std::string &root);
+        VoxCeleb(const std::string &root, DataMode mode);
+        VoxCeleb(const std::string &root, DataMode mode , bool download);
+        VoxCeleb(const std::string &root, DataMode mode , bool download, TransformType transforms);
+
+        private :
+            void load_data();
+
+        void check_resources();
+    };
+}
