+++ src/datasets/audio-speech/vox-celeb.cpp
@@ -1,21 +1,21 @@
-#include "../../../include/datasets/audio-speech/cmu-arctic.h"
+#include "../../../include/datasets/audio-speech/vox-celeb.h"
 
 namespace xt::data::datasets {
 
-    CMUArctic::CMUArctic(const std::string &root): CMUArctic::CMUArctic(root, DataMode::TRAIN, false) {
+    VoxCeleb::VoxCeleb(const std::string &root): VoxCeleb::VoxCeleb(root, DataMode::TRAIN, false) {
     }
 
-    CMUArctic::CMUArctic(const std::string &root, DataMode mode): CMUArctic::CMUArctic(root, mode, false) {
+    VoxCeleb::VoxCeleb(const std::string &root, DataMode mode): VoxCeleb::VoxCeleb(root, mode, false) {
     }
 
-    CMUArctic::CMUArctic(const std::string &root, DataMode mode, bool download) : BaseDataset(root, mode, download) {
-        throw std::runtime_error("CMUArctic: CMUArctic not implemented");
+    VoxCeleb::VoxCeleb(const std::string &root, DataMode mode, bool download) : BaseDataset(root, mode, download) {
+        throw std::runtime_error("VoxCeleb: VoxCeleb not implemented");
     }
 
 
-    CMUArctic::CMUArctic(const std::string &root, DataMode mode, bool download,
+    VoxCeleb::VoxCeleb(const std::string &root, DataMode mode, bool download,
                            TransformType transforms) : BaseDataset(root, mode, download, transforms) {
-        throw std::runtime_error("CMUArctic: CMUArctic not implemented");
+        throw std::runtime_error("VoxCeleb: VoxCeleb not implemented");
     }
 
 
