+++ include/datasets/language-modeling/en-wik.h
@@ -1,8 +1,19 @@
-//
-// Created by kami on 4/15/25.
-//
+#pragma once
+#include "../base/base.h"
+#include "../../headers/datasets.h"
 
-#ifndef EN_WIK_H
-#define EN_WIK_H
 
-#endif //EN_WIK_H
+namespace xt::data::datasets {
+    class EnWik : BaseDataset {
+        public :
+            explicit EnWik(const std::string &root);
+        EnWik(const std::string &root, DataMode mode);
+        EnWik(const std::string &root, DataMode mode , bool download);
+        EnWik(const std::string &root, DataMode mode , bool download, TransformType transforms);
+
+        private :
+            void load_data();
+
+        void check_resources();
+    };
+}
