+++ src/datasets/language-modeling/en-wik.cpp
@@ -1,3 +1,0 @@
-//
-// Created by kami on 4/15/25.
-//
