+++ include/datasets/language-modeling/penn-treebank.h
@@ -1,8 +1,19 @@
-//
-// Created by kami on 4/15/25.
-//
+#pragma once
+#include "../base/base.h"
+#include "../../headers/datasets.h"
 
-#ifndef PENN_TREEBANK_H
-#define PENN_TREEBANK_H
 
-#endif //PENN_TREEBANK_H
+namespace xt::data::datasets {
+    class PennTreebank : BaseDataset {
+        public :
+            explicit PennTreebank(const std::string &root);
+        PennTreebank(const std::string &root, DataMode mode);
+        PennTreebank(const std::string &root, DataMode mode , bool download);
+        PennTreebank(const std::string &root, DataMode mode , bool download, TransformType transforms);
+
+        private :
+            void load_data();
+
+        void check_resources();
+    };
+}
