+++ src/datasets/language-modeling/penn-treebank.cpp
@@ -1,3 +1,22 @@
-//
-// Created by kami on 4/15/25.
-//
+#include "../../../include/datasets/language-modeling/penn-treebank.h"
+
+namespace xt::data::datasets {
+
+    PennTreebank::PennTreebank(const std::string &root): PennTreebank::PennTreebank(root, DataMode::TRAIN, false) {
+    }
+
+    PennTreebank::PennTreebank(const std::string &root, DataMode mode): PennTreebank::PennTreebank(root, mode, false) {
+    }
+
+    PennTreebank::PennTreebank(const std::string &root, DataMode mode, bool download) : BaseDataset(root, mode, download) {
+        throw std::runtime_error("PennTreebank: PennTreebank not implemented");
+    }
+
+
+    PennTreebank::PennTreebank(const std::string &root, DataMode mode, bool download,
+                           TransformType transforms) : BaseDataset(root, mode, download, transforms) {
+        throw std::runtime_error("PennTreebank: PennTreebank not implemented");
+    }
+
+
+}
