+++ include/datasets/language-modeling/wiki-text.h
@@ -1,8 +1,19 @@
-//
-// Created by kami on 4/15/25.
-//
+#pragma once
+#include "../base/base.h"
+#include "../../headers/datasets.h"
 
-#ifndef WIKI_TEXT_H
-#define WIKI_TEXT_H
 
-#endif //WIKI_TEXT_H
+namespace xt::data::datasets {
+    class WikiText : BaseDataset {
+        public :
+            explicit WikiText(const std::string &root);
+        WikiText(const std::string &root, DataMode mode);
+        WikiText(const std::string &root, DataMode mode , bool download);
+        WikiText(const std::string &root, DataMode mode , bool download, TransformType transforms);
+
+        private :
+            void load_data();
+
+        void check_resources();
+    };
+}
