+++ include/datasets/machine-translation/iwslt.h
@@ -1,8 +1,19 @@
-//
-// Created by kami on 4/15/25.
-//
+#pragma once
+#include "../base/base.h"
+#include "../../headers/datasets.h"
 
-#ifndef IWSLT_H
-#define IWSLT_H
 
-#endif //IWSLT_H
+namespace xt::data::datasets {
+    class IWSLT : BaseDataset {
+        public :
+            explicit IWSLT(const std::string &root);
+        IWSLT(const std::string &root, DataMode mode);
+        IWSLT(const std::string &root, DataMode mode , bool download);
+        IWSLT(const std::string &root, DataMode mode , bool download, TransformType transforms);
+
+        private :
+            void load_data();
+
+        void check_resources();
+    };
+}
