+++ include/datasets/machine-translation/multi.h
@@ -1,8 +1,19 @@
-//
-// Created by kami on 4/15/25.
-//
+#pragma once
+#include "../base/base.h"
+#include "../../headers/datasets.h"
 
-#ifndef MULTI_H
-#define MULTI_H
 
-#endif //MULTI_H
+namespace xt::data::datasets {
+    class MULTI : BaseDataset {
+        public :
+            explicit MULTI(const std::string &root);
+        MULTI(const std::string &root, DataMode mode);
+        MULTI(const std::string &root, DataMode mode , bool download);
+        MULTI(const std::string &root, DataMode mode , bool download, TransformType transforms);
+
+        private :
+            void load_data();
+
+        void check_resources();
+    };
+}
