+++ include/datasets/machine-translation/wmt.h
@@ -1,8 +1,19 @@
-//
-// Created by kami on 4/15/25.
-//
+#pragma once
+#include "../base/base.h"
+#include "../../headers/datasets.h"
 
-#ifndef WMT_H
-#define WMT_H
 
-#endif //WMT_H
+namespace xt::data::datasets {
+    class WMT : BaseDataset {
+        public :
+            explicit WMT(const std::string &root);
+        WMT(const std::string &root, DataMode mode);
+        WMT(const std::string &root, DataMode mode , bool download);
+        WMT(const std::string &root, DataMode mode , bool download, TransformType transforms);
+
+        private :
+            void load_data();
+
+        void check_resources();
+    };
+}
