+++ include/datasets/natural-language-inference/cola.h
@@ -1,8 +1,19 @@
-//
-// Created by kami on 4/15/25.
-//
+#pragma once
+#include "../base/base.h"
+#include "../../headers/datasets.h"
 
-#ifndef COLA_H
-#define COLA_H
 
-#endif //COLA_H
+namespace xt::data::datasets {
+    class COLA : BaseDataset {
+        public :
+            explicit COLA(const std::string &root);
+        COLA(const std::string &root, DataMode mode);
+        COLA(const std::string &root, DataMode mode , bool download);
+        COLA(const std::string &root, DataMode mode , bool download, TransformType transforms);
+
+        private :
+            void load_data();
+
+        void check_resources();
+    };
+}
