+++ include/datasets/natural-language-inference/mnli.h
@@ -1,8 +1,19 @@
-//
-// Created by kami on 4/15/25.
-//
+#pragma once
+#include "../base/base.h"
+#include "../../headers/datasets.h"
 
-#ifndef MNLI_H
-#define MNLI_H
 
-#endif //MNLI_H
+namespace xt::data::datasets {
+    class MNLI : BaseDataset {
+        public :
+            explicit MNLI(const std::string &root);
+        MNLI(const std::string &root, DataMode mode);
+        MNLI(const std::string &root, DataMode mode , bool download);
+        MNLI(const std::string &root, DataMode mode , bool download, TransformType transforms);
+
+        private :
+            void load_data();
+
+        void check_resources();
+    };
+}
