+++ include/datasets/natural-language-inference/mrpc.h
@@ -1,8 +1,19 @@
-//
-// Created by kami on 4/15/25.
-//
+#pragma once
+#include "../base/base.h"
+#include "../../headers/datasets.h"
 
-#ifndef MRPC_H
-#define MRPC_H
 
-#endif //MRPC_H
+namespace xt::data::datasets {
+    class MRPC : BaseDataset {
+        public :
+            explicit MRPC(const std::string &root);
+        MRPC(const std::string &root, DataMode mode);
+        MRPC(const std::string &root, DataMode mode , bool download);
+        MRPC(const std::string &root, DataMode mode , bool download, TransformType transforms);
+
+        private :
+            void load_data();
+
+        void check_resources();
+    };
+}
