### 2025-07-10 #8565 @editorialbot
[REVIEW]: xTorch: A High-Level C++ Extension Library for PyTorch (LibTorch)
**Submitting author:** @kamisaberi (<a href="http://orcid.org/0009-0002-7822-6168">kamran saberifard</a>)
**Repository:** https://github.com/kamisaberi/xtorch
**Branch with paper.md** (empty if default branch): main
**Version:** v0.2.0
