### 2025-07-16 kamisaberi/xtorch#956 @justusschock
[JOSS REVIEW] Missing Tests
Hey @kamisaberi , while this is a great piece of software, for the JOSS paper over at https://github.com/openjournals/joss-reviews/issues/8565 , it is required for this lib to have a fully automated test suite covering at least the core of the functionality.
