### 2025-07-16 kamisaberi/xtorch#957 @justusschock
[JOSS REVIEW] Missing Documentation
Hey @kamisaberi , in addition to the examples, a API reference (e.g. one generated with readthedocs (which you seem to have configured) or doxygen would be required for proper documentation (see checklists from https://github.com/openjournals/joss-reviews/issues/8565) . Also an introduction/install/quickstart beyond the readme would be great.
