### 2025-07-16 kamisaberi/xtorch#958 @justusschock
[JOSS REVIEW] Performance Claims
For the joss review over in https://github.com/openjournals/joss-reviews/issues/8565 we need to replicate the performance claims.

In your Readme you make the following claims:
> XTorch on a single NVIDIA RTX 3090 GPU completed a 5-epoch run in 219 seconds. This represents a 37.4% speedup over a standard PyTorch implementation which required 350 seconds using two RTX 3090 GPUs with DataParallel

This isn't a fair comparison since DataParallel is known to be bad for performance. Could you provide benchmarks with single GPU (same number of loading threads) for both and (optionally) benchmark DDP with two gpus and same configuration each?

Additionally, could you provide more machine specs like num cpus, amount of RAM, os etc?
