### 2025-08-07 #8565 @justusschock
@vissarion , could you advice how to proceed here please? without these issues addressed to some extend, I cannot recommend this paper for publication.
