### 2026-04-29 #8565 @vissarion
> [@editorialbot](https://github.com/editorialbot) remind me in 1 day
>
> [@vissarion](https://github.com/vissarion) I think we are ready to reject. I can do this tomorrow after you confirm. Thank you.

Hi @crvernon, I haven't realized you are waiting for my confirmation. I confirm that we should reject. If @kamisaberi is interested and have time in the future can submit again.
